make jet field's compatible with main branch

+++ api/v1alpha1/hazelcast_types.go
@@ -219,22 +219,22 @@ type JetEngineConfiguration struct {
 type JetInstance struct {
 	// The number of threads Jet creates in its cooperative multithreading pool.
 	// +optional
-	CooperativeThreadCount int `json:"cooperativeThreadCount,omitempty"`
+	CooperativeThreadCount int32 `json:"cooperativeThreadCount,omitempty"`
 
 	// The duration of the interval between flow-control packets.
 	// +kubebuilder:default:=100
 	// +optional
-	FlowControlPeriodMillis int `json:"flowControlPeriodMillis,omitempty"`
+	FlowControlPeriodMillis int32 `json:"flowControlPeriodMillis,omitempty"`
 
 	// The number of synchronous backups to configure on the IMap that Jet needs internally to store job metadata and snapshots.
 	// +kubebuilder:default:=1
 	// +optional
-	BackupCount int `json:"backupCount,omitempty"`
+	BackupCount int32 `json:"backupCount,omitempty"`
 
 	// The delay after which the auto-scaled jobs restart if a new member joins the cluster.
 	// +kubebuilder:default:=10000
 	// +optional
-	ScaleUpDelayMillis int `json:"scaleUpDelayMillis,omitempty"`
+	ScaleUpDelayMillis int32 `json:"scaleUpDelayMillis,omitempty"`
 
 	// Specifies whether the Lossless Cluster Restart feature is enabled.
 	// +kubebuilder:default:=false
@@ -243,7 +243,7 @@ type JetInstance struct {
 
 	// Specifies the maximum number of records that can be accumulated by any single processor instance.
 	// +optional
-	MaxProcessorAccumulatedRecords int `json:"maxProcessorAccumulatedRecords,omitempty"`
+	MaxProcessorAccumulatedRecords int64 `json:"maxProcessorAccumulatedRecords,omitempty"`
 }
 
 // Returns true if Jet Instance section is configured.
@@ -254,15 +254,15 @@ func (j *JetInstance) IsConfigured() bool {
 type JetEdgeDefaults struct {
 	// Sets the capacity of processor-to-processor concurrent queues.
 	// +optional
-	QueueSize int `json:"queueSize,omitempty"`
+	QueueSize int32 `json:"queueSize,omitempty"`
 
 	// Limits the size of the packet in bytes.
 	// +optional
-	PacketSizeLimit int `json:"packetSizeLimit,omitempty"`
+	PacketSizeLimit int32 `json:"packetSizeLimit,omitempty"`
 
 	// Sets the scaling factor used by the adaptive receive window sizing function.
 	// +optional
-	ReceiveWindowMultiplier int `json:"receiveWindowMultiplier,omitempty"`
+	ReceiveWindowMultiplier int8 `json:"receiveWindowMultiplier,omitempty"`
 }
 
 // Returns true if Jet Instance Edge Defaults is configured.
